configurador/requirements.txt
fastapi
httpx
uvicorn
pyyaml 
pydantic
jinja2
pyOpenSSL
cryptography
itsdangerous
python-multipart
passlib
